Convert spaces to tabs and extract MavenProjectBuilder test class to its own file
