+++ src/test/java/com/mfoot/mojo/libyear/InMemoryTestLogger.java
@@ -23,87 +23,88 @@ import java.util.List;
 
 class InMemoryTestLogger implements Log {
 
-    public final List<String> infoLogs = new ArrayList<>();
-    public final List<String> errorLogs = new ArrayList<>();
-    public final List<String> debugLogs = new ArrayList<>();
+	public final List<String> infoLogs = new ArrayList<>();
+	public final List<String> errorLogs = new ArrayList<>();
+	public final List<String> debugLogs = new ArrayList<>();
 
-    @Override
-    public boolean isDebugEnabled() {
-        return true;
-    }
+	@Override
+	public boolean isDebugEnabled() {
+		return true;
+	}
 
-    @Override
-    public void debug(CharSequence charSequence) {
-        debugLogs.add(charSequence.toString());
-    }
+	@Override
+	public void debug(CharSequence charSequence) {
+		System.out.println(charSequence);
+		debugLogs.add(charSequence.toString());
+	}
 
-    @Override
-    public void debug(CharSequence charSequence, Throwable throwable) {
+	@Override
+	public void debug(CharSequence charSequence, Throwable throwable) {
 
-    }
+	}
 
-    @Override
-    public void debug(Throwable throwable) {
+	@Override
+	public void debug(Throwable throwable) {
 
-    }
+	}
 
-    @Override
-    public boolean isInfoEnabled() {
-        return false;
-    }
+	@Override
+	public boolean isInfoEnabled() {
+		return false;
+	}
 
-    @Override
-    public void info(CharSequence charSequence) {
-        infoLogs.add(charSequence.toString());
-    }
+	@Override
+	public void info(CharSequence charSequence) {
+		infoLogs.add(charSequence.toString());
+	}
 
-    @Override
-    public void info(CharSequence charSequence, Throwable throwable) {
+	@Override
+	public void info(CharSequence charSequence, Throwable throwable) {
 
-    }
+	}
 
-    @Override
-    public void info(Throwable throwable) {
+	@Override
+	public void info(Throwable throwable) {
 
-    }
+	}
 
-    @Override
-    public boolean isWarnEnabled() {
-        return false;
-    }
+	@Override
+	public boolean isWarnEnabled() {
+		return false;
+	}
 
-    @Override
-    public void warn(CharSequence charSequence) {
+	@Override
+	public void warn(CharSequence charSequence) {
 
-    }
+	}
 
-    @Override
-    public void warn(CharSequence charSequence, Throwable throwable) {
+	@Override
+	public void warn(CharSequence charSequence, Throwable throwable) {
 
-    }
+	}
 
-    @Override
-    public void warn(Throwable throwable) {
+	@Override
+	public void warn(Throwable throwable) {
 
-    }
+	}
 
-    @Override
-    public boolean isErrorEnabled() {
-        return false;
-    }
+	@Override
+	public boolean isErrorEnabled() {
+		return false;
+	}
 
-    @Override
-    public void error(CharSequence charSequence) {
-        errorLogs.add(charSequence.toString());
-    }
+	@Override
+	public void error(CharSequence charSequence) {
+		errorLogs.add(charSequence.toString());
+	}
 
-    @Override
-    public void error(CharSequence charSequence, Throwable throwable) {
+	@Override
+	public void error(CharSequence charSequence, Throwable throwable) {
 
-    }
+	}
 
-    @Override
-    public void error(Throwable throwable) {
+	@Override
+	public void error(Throwable throwable) {
 
-    }
+	}
 }
